bug fixed

+++ kernel_reinstall.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 . ./common.sh --source-only
 ENV_TARGET=env_kernel.rc
+myAnswer=n
 source $ENV_TARGET &&
     echo -e "[PHW]ENV sourced ${myGR}$ENV_TARGET${myNC}"
 CUR_DIR=`pwd`
@@ -14,9 +15,9 @@ cd $KERNEL_HOME &&
 # cd $KERNEL_ORIGIN &&
 #     echo -e "[PHW]move dir into ${myBL}$KERNEL_ORIGIN${myNC}"
 
-make -j8 | tee build.log &&
-    make modules_install -j8 | tee module_install.log &&
-    make install -j8 | tee install.log&&
+make -j6 | tee build.log &&
+    make modules_install -j6 | tee module_install.log &&
+    make install -j6 | tee install.log&&
     echo -e "[PHW]${myGR}SUCCESS${myNC} to kernel install" &&
     update-grub2 > /dev/null &&
     echo -e "[PHW]${myGR}SUCCESS${myNC} to grub2-update"
@@ -25,6 +26,9 @@ if [[ $? != 0 ]]; then
     echo -e "[PHW]${myRED}FAIL${myNC} to kernel install" &&
         cd $CUR_DIR
 else
-    [ "$myAnswer" == "y" ] && echo -e "After 5 second, reboot will proceed"
-    reboot
+    if [ "$myAnswer" == "y" ]; then
+        echo -e "After 5 second, reboot will proceed"
+        sleep 5
+        reboot
+    fi
 fi
